First (final) submission for Sudoku puzzle solving Challenge. code version-5. 22.05.2023
This code uses- constant propagation and backtracking method to solve Sudoku puzzles. It doesn't use grid format of Sudoku, it solves Sudoku in it's raw string format itself. It uses dictionaries of positions named- units and peers, and makes a dictionary of input puzzle with respective positions in Sudoku grid. Then applies the constant propagation and backtracking method to solve puzzle. The current code solves one puzzle at one time, and then iterates to other (Here, there is chance of improvement to make the code more faster by assigning system to solve multiple puzzles at one time). However, this code is embedded with the additional code that allows this code to use multiple cores of a CPU and do multi-threading. It significantly reduces the task-time by as much number of folds as there are cores in the CPU on which code is being run. It takes 0.4 seconds to solve 1000 easy level Sudoku problems, 1.18 seconds for 100 hard level problems, 3.0 seconds for 1000 harder level problems- in a 8 core CPU. It takes input file in CSV format (e.g. Sudoku.csv). Enjoy the speed. Thank you, Arvind.